Refactor into event submodule
