+++ proxy/core/event/__init__.py
@@ -1,0 +1,22 @@
+# -*- coding: utf-8 -*-
+"""
+    proxy.py
+    ~~~~~~~~
+    ⚡⚡⚡ Fast, Lightweight, Pluggable, TLS interception capable proxy server focused on
+    Network monitoring, controls & Application development, testing, debugging.
+
+    :copyright: (c) 2013-present by Abhinav Singh and contributors.
+    :license: BSD, see LICENSE for more details.
+"""
+from .queue import EventQueue
+from .names import EventNames, eventNames
+from .dispatcher import EventDispatcher
+from .subscriber import EventSubscriber
+
+__all__ = [
+    'eventNames',
+    'EventNames',
+    'EventQueue',
+    'EventDispatcher',
+    'EventSubscriber',
+]
